pyproject.toml
[project]
name = "horosc-hope"
version = "0.1.0"
description = "A horoscope application with AI-powered features"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "requests",
    "beautifulsoup4",
    "openai",
    "python-dotenv",
    "ephem>=4.2",
    "pyswisseph>=2.10.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "flake8",
    "mypy",
] 
